feat: add unregister invocation and use spin lock to ensure thread safety
